+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Optional prompt template for skills
 - Optional prompt hint for skills
 
 ## v0.0.8 - 2026-07-05